pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "copinance-os"
version = "0.1.0"
description = "Open-source market analysis platform and framework with question-driven AI and deterministic instrument/macro analysis"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Copinance OS Contributors"}
]
keywords = ["market", "research", "finance", "trading", "options", "ai", "analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License 2.0",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.2",
    "pandas>=2.3.3",
    "numpy>=1.26.0,<3.0",
    "structlog>=25.5.0",
    "typer>=0.20.1",
    "rich>=14.2.0",
    "yfinance>=0.2.66",
    "google-genai>=1.56.0",
    "httpx>=0.27.0",
    "dependency-injector>=4.48.3",
    "QuantLib>=1.41",
    "edgartools==5.26.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "black>=25.12.0",
    "ruff>=0.14.10",
    "mypy>=1.19.1",
    "pandas-stubs>=2.2.0",
    "types-PyYAML>=6.0.12",
    "pre-commit>=4.5.1",
    "fastapi>=0.115.0",
    "pyyaml>=6.0.1",
]


ollama = [
    "httpx>=0.27.0",
]

api = [
    "fastapi>=0.115.0",
]

plugins = [
    "pyyaml>=6.0.1",
]

[project.scripts]
copinance = "copinance_os.interfaces.cli:app"

[project.urls]
Homepage = "https://github.com/copinance/copinance-os"
Documentation = "https://copinance.github.io/copinance-os/"
Repository = "https://github.com/copinance/copinance-os"
Issues = "https://github.com/copinance/copinance-os/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["copinance_os*"]

[tool.setuptools.package-data]
copinance_os = [
    "py.typed",
    "ai/llm/resources/*.json",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=copinance_os",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
filterwarnings = [
    # Suppress ResourceWarnings from yfinance/pandas/numpy internal SQLite caching
    # These are from third-party libraries and not actionable for our code
    "ignore::ResourceWarning",
    "ignore:unclosed.*:ResourceWarning",
    "ignore:.*unclosed.*:ResourceWarning",
    # Suppress event loop deprecation warnings in test fixtures
    "ignore:There is no current event loop:DeprecationWarning",
    # Third-party: google-genai uses _UnionGenericAlias (deprecated in 3.14, removal 3.17)
    "ignore:_UnionGenericAlias is deprecated:DeprecationWarning",
    "ignore:.*_UnionGenericAlias.*:DeprecationWarning",
    # Unittest mock: unawaited coroutine when AsyncMock is used as return value and code does not await
    "ignore:coroutine.*was never awaited:RuntimeWarning",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",        # pycodestyle errors (includes E402: module level import not at top of file)
    "W",        # pycodestyle warnings
    "F",        # pyflakes
    "I",        # isort
    "B",        # flake8-bugbear
    "C4",       # flake8-comprehensions
    "UP",       # pyupgrade
    "N",        # pep8-naming: enforces Python naming conventions
    "PLC0415",  # import-outside-top-level: checks imports are not inside functions/blocks
    "D1",       # pydocstyle: docstring conventions (D100 module, D107 __init__, etc.)
    "SIM",      # flake8-simplify: simplify code (e.g. if/else, ternary)
    "PTH",      # flake8-use-pathlib: use pathlib instead of os.path
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "E402",  # module level import not at top of file (allow for optional imports in try/except blocks)
]
# D1 (docstrings): require module docstrings (D100); method/class docstrings (D101, D102, D103, D107) encouraged but not enforced to avoid bulk churn
[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D100", "D101", "D102", "D103", "D107"]
".pre-commit-hooks/*" = ["D100", "D101", "D102", "D103", "D107"]
"src/**" = ["D101", "D102", "D103", "D107"]
# Note: PLC0415 (import-outside-top-level) will flag imports inside functions/methods.
# This helps enforce PEP 8 style: imports should be at module level, not inside functions.
# Imports in try/except blocks at module level are acceptable for optional dependencies.

[tool.ruff.lint.isort]
known-first-party = ["copinance_os"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
strict_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = [
    "httpx",
    "httpx.*",
]
ignore_missing_imports = true

# edgartools exposes `edgar` with no py.typed; import-untyped is raised in our provider file.
[[tool.mypy.overrides]]
module = "copinance_os.data.providers.sec.edgartools"
disable_error_code = ["import-untyped"]

# QuantLib publishes no types / py.typed; local mypy reports import-untyped. Pre-commit
# mypy runs without QuantLib installed, so an inline ignore there becomes unused-ignore.
[[tool.mypy.overrides]]
module = "QuantLib"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "copinance_os.ai.llm.providers.ollama",
    "copinance_os.ai.llm.providers.gemini",
]
disable_error_code = ["unused-ignore"]

[[tool.mypy.overrides]]
module = [
    "copinance_os.data.providers.yfinance",
    "copinance_os.research.workflows.market",
]
disable_error_code = ["unused-ignore"]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
